Invoice tracking Bakend/requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
astor==0.8.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cron_descriptor==2.0.6
cryptography==45.0.7
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
Django==5.2.5
django-auditlog==3.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
django-seed @ file:///C:/Users/Administrateur/Documents/GitHub/django-seed
django-timezone-field==7.1
django-utils==0.0.2
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
drf-spectacular==0.28.0
einops==0.8.1
environ==1.0
environs==14.3.0
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==37.8.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
idna==3.10
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-text-splitters==0.3.8
langsmith==0.4.2
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
MouseInfo==0.1.3
multidict==6.5.1
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.42.1
networkx==3.5
numpy==2.3.0
openai==1.91.0
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.16.0
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
paddleocr==3.0.2
paddlepaddle==3.0.0
paddlex==3.0.2
pandas==2.3.0
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.2.1
pipreqs==0.4.13
platformdirs==4.4.0
pluggy==1.6.0
premailer==3.10.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
pygame-ce==2.5.5
pygame_gui==0.6.14
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.2
pyparsing==3.2.3
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-i18n==0.3.9
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
selenium==4.33.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
toposort==1.10
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarg==0.1.10
yarl==1.20.1
zstandard==0.23.0
